adidas-ecommerce/services/payment-service/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.7
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-allauth>=0.57.0
celery>=5.3.0
redis>=5.0.0
django-extensions>=3.2.0
gunicorn>=21.2.0
whitenoise>=6.6.0
